+++ pkg/paxos/message_test.go
@@ -158,9 +158,9 @@ func TestNewVoteAlt(t *testing.T) {
 func TestSetFrom(t *testing.T) {
     m := NewInvite(1)
     m.SetFrom(1)
-    assert.Equal(t, uint64(1), m.From(), "")
+    assert.Equal(t, 1, m.From(), "")
     m.SetFrom(2)
-    assert.Equal(t, uint64(2), m.From(), "")
+    assert.Equal(t, 2, m.From(), "")
 }
 
 func TestSetSeqn(t *testing.T) {
